requirements.txt
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
chromadb
wikipedia
arxiv
faiss-cpu
groq
langchain-groq
beautifulsoup4
cassio